refactor: market info composite
