+++ apps/main/src/lend/components/PageLendMarket/LendMarketPage.tsx
@@ -7,7 +7,7 @@ import { useLendPageTitle } from '@/lend/hooks/useLendPageTitle'
 import { networks } from '@/lend/networks'
 import { type MarketUrlParams } from '@/lend/types/lend.types'
 import { getCollateralListPathname, parseMarketParams } from '@/lend/utils/utilsRouter'
-import { MarketOverviewCard } from '@/llamalend/features/market-advanced-information'
+import { MarketOverviewCard } from '@/llamalend/features/market-advanced-information/MarketOverviewCard'
 import { MarketContextProvider } from '@/llamalend/features/market-context'
 import { PositionDetailsComposite } from '@/llamalend/features/market-position-details'
 import { useIsInLiquidation } from '@/llamalend/features/market-position-details/hooks/useUserLiquidationStatus'
@@ -111,11 +111,7 @@ export const LendMarketPage = () => {
         />
         <PositionDetailsComposite hasPosition={loanExists} events={collateralEvents} />
         {isMarketDetailPageV2 && <MarketOverviewCard />}
-        <MarketInformationComposite
-          rateType={MarketRateType.Borrow}
-          previewPrices={previewPrices}
-          isMarketDetailPageV2={isMarketDetailPageV2}
-        />
+        <MarketInformationComposite rateType={MarketRateType.Borrow} previewPrices={previewPrices} />
       </DetailPageLayout>
     </MarketContextProvider>
   )
