+++ apps/main/src/llamalend/features/market-advanced-information/index.ts
@@ -3,4 +3,4 @@ export * from './MarketContractsSection'
 export * from './MarketLoanParameters'
 export * from './MarketParametersSection'
 export * from './AdvancedDetails'
-export * from './MarketDetailsCard'
+export * from './MarketParametersCard'
