+++ apps/main/src/loan/components/MarketInformationComposite.tsx
@@ -13,6 +13,7 @@ import CardContent from '@mui/material/CardContent'
 import CardHeader from '@mui/material/CardHeader'
 import Stack from '@mui/material/Stack'
 import type { Decimal } from '@primitives/decimal.utils'
+import { useLlamaMarketDetailPageV2 } from '@ui-kit/hooks/useFeatureFlags'
 import { t } from '@ui-kit/lib/i18n'
 import { MarketRateType } from '@ui-kit/types/market'
 import type { Range } from '@ui-kit/types/util'
@@ -22,11 +23,11 @@ import { networks } from '../networks'
 
 type MarketInformationCompProps = {
   previewPrices: Range<Decimal> | undefined
-  isMarketDetailPageV2: boolean
 }
 
-export const MarketInformationComposite = ({ previewPrices, isMarketDetailPageV2 }: MarketInformationCompProps) => {
+export const MarketInformationComposite = ({ previewPrices }: MarketInformationCompProps) => {
   const { chainId } = useMarketContext<ChainId>()
+  const isMarketDetailPageV2 = useLlamaMarketDetailPageV2()
   return (
     <Stack sx={{ gap: PAGE_SPACING }}>
       <ChartAndActivityComp previewPrices={previewPrices} />
